refactoring duplicate code
